240108 Controller들 @RequestParam 변수 수정 & try-catch 구문 추가

+++ src/main/java/com/backend/jejutour_server/repository/TourRepository.java
@@ -10,7 +10,7 @@ import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface TourRepository extends JpaRepository<TourEntity,Integer>, TourRepositoryCustom {
+public interface TourRepository extends JpaRepository<TourEntity, Long >, TourRepositoryCustom {
 
 
     List<TourEntity> findByItemsRegion2CdValue(Long region2cd_label);
